refactor: rename lib folder liblogosdelivery -> library
Keep the directory named `library` (the file/lib/header stay
`liblogosdelivery.*`) so consumers import
`logos-delivery/library/liblogosdelivery.h` instead of the redundant
`liblogosdelivery/liblogosdelivery.h`.

Why: addresses PR review — the doubled path segment read awkwardly;
a single `library/` dir exposing `library/liblogosdelivery.h` is cleaner.

Updates the folder path everywhere it is referenced: kernel_api
absolute imports, Makefile/nimble build targets and iOS stubs, the
example include paths and the RN build script.
