+++ library/declare_lib.nim
