+++ library/examples/logosdelivery_example.c
