+++ library/ios_natpmp_stubs.c
