+++ library/kernel_api/debug_node_api.nim
@@ -12,7 +12,7 @@ import
   logos_delivery/waku/factory/waku,
   logos_delivery/waku/node/waku_node,
   logos_delivery/waku/node/health_monitor,
-  liblogosdelivery/declare_lib
+  library/declare_lib
 
 proc getMultiaddresses(node: WakuNode): seq[string] =
   return node.info().listenAddresses
