+++ library/liblogosdelivery.h
