+++ library/nim.cfg
